refactor: reissue service
